+++ src/importlinter/adapters/user_options.py
@@ -51,7 +51,9 @@ class IniFileUserOptionReader(AbstractUserOptionReader):
     section_name = "importlinter"
 
     def _read_config_filename(self, config_filename: str) -> Optional[UserOptions]:
-        config = configparser.ConfigParser()
+        config = configparser.ConfigParser(
+            interpolation=configparser.ExtendedInterpolation(),
+        )
         file_contents = settings.FILE_SYSTEM.read(config_filename)
         config.read_string(file_contents)
         if self.section_name in config.sections():